+++ NEWS.md
@@ -5,6 +5,7 @@ Bug fixes
 ---------
 
  * `confint.profile.gnm()` now works for a single parameter  ![(#10)](https://github.com/hturner/gnm/issues/10)
+  * `gnm()` now works when `eliminate` is specified as `NULL` ![(#14)](https://github.com/hturner/gnm/issues/14)
  
 Changes in gnm 1.1-0
 ====================